Add code from Spring Boot live demo
* refactored and commented for assignment 4
