+++ application/src/main/java/de/unibayreuth/se/teaching/list/Application.java
@@ -1,0 +1,14 @@
+package de.unibayreuth.se.teaching.list;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Main class to start the Spring Boot application in production.
+ */
+@SpringBootApplication
+public class Application {
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
