+++ application/src/main/resources/application.yaml
@@ -1,0 +1,15 @@
+spring:
+  application:
+    name: se24-demo
+  #datasource:
+  #  driver-class-name: org.postgresql.Driver
+  #flyway:
+  #  enabled: true
+  #  locations: classpath:db/migration
+  #  validate-on-migrate: true
+
+#springdoc:
+#    api-docs:
+#        path: /api-docs
+#    swagger-ui:
+#        path: /swagger-ui.html
