+++ lombok.config
@@ -1,0 +1,4 @@
+# Copy the Qualifier annotation from the instance variables to the constructor
+# see https://github.com/rzwitserloot/lombok/issues/745
+lombok.copyableAnnotations += org.springframework.beans.factory.annotation.Qualifier
+lombok.copyableAnnotations += org.springframework.beans.factory.annotation.Value
