--- src/main/java/de/unibayreuth/se/teaching/list/Application.java
@@ -1,11 +1,0 @@
-package de.unibayreuth.se.teaching.list;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
