--- src/main/resources/application.yaml
@@ -1,3 +1,0 @@
-spring:
-  application:
-    name: se24-linked-list
